+++ CHANGELOG.md
@@ -6,17 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
-- Fix local package build process (#488)
+### Added
+
+- New colors (#474)
 
 ### Changed
 
 - Updated spacing variables to match the new design system (#460)
+- Color usage (#474)
 
 ### Fixed
 
 - Modal button alignment support (#460)
 - Storybook setup and dev deployment (#461)
 - Updating Sentry to fix compile-time error (#487)
+- Fix local package build process (#488)
+- Color naming convention (#474)
 
 ## [0.15.0] - 2023-04-27