+++ CHANGELOG.md
@@ -10,8 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 - Ability to create `HTML` and `CSS` files in a `HTML` project (#478)
 - Help text in the new file modal (#478)
+
 ### Changed
 
+- Handling of internal/external links for HTML/CSS projects (#483)
 - Position run button in separate bar (#471)
 
 ### Fixed