+++ CHANGELOG.md
@@ -6,6 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Added
+
+- New colors (#474)
+
+### Changed
+
+- Updated spacing variables to match the new design system (#460)
+- Color usage (#474)
+
+### Fixed
+
+- Modal button alignment support (#460)
+- Storybook setup and dev deployment (#461)
+- Updating Sentry to fix compile-time error (#487)
+- Fix local package build process (#488)
+- Color naming convention (#474)
 
 ## [0.15.0] - 2023-04-27
 
@@ -18,12 +34,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Changed
 
+- Projects tab layout (#447)
 - Refactored `NewComponentButton` (#450)
 - Modal styling (#450)
-- Projects tab layout (#447)
 
 ### Fixed
 
+- Fix storybook build error (#484)
 - CookieBot consent request in iframes (#437)
 - Resetting name error if file creation/renaming is cancelled (#450)
 - Mission Zero performance gains (#457)
@@ -33,6 +50,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Container queries (#445)
+- Resize handle #453
 - File type icons (#449)
 
 ## [0.14.0] - 2023-03-28