+++ CHANGELOG.md
@@ -19,15 +19,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Fixed
 
 - Updated Sentry to fix compile-time error
+- Ability to run HTML projects by clicking run button (#481)
 
 ### Changed
 
 - Updated spacing variables to match the new design system (#460)
+- Position run button in separate bar (#471)
 
 ### Fixed
 
 - Modal button alignment support (#460)
 - Storybook setup and dev deployment (#461)
+- HTML auto-run functionality (#481)
 - Updating Sentry to fix compile-time error (#487)
 
 ## [0.15.0] - 2023-04-27